Move search forms under widgets
